mediclear-backend/requirements.txt
##requirements.py
streamlit
openai
Pillow
pytesseract
python-dotenv
pymongo
